FashionTagger/requirements.txt
tensorflow==2.19.0
flask==2.2.5
pillow==9.5.0
